requirements.txt
flask-restx==1.3.0
pymongo==4.10.1
tzdata==2025.1
factory_boy==3.3.3
pytest==8.3.4
python-dotenv==1.0.1
pytest-cov==7.0.0
gunicorn==25.1.0